+++ src/cargo.rs
@@ -144,7 +144,13 @@ pub(crate) fn check(project: &Project, test: &Test, options: &Options) -> Result
 pub(crate) fn test(project: &Project, test: &Test, options: &Options) -> Result<CargoOutput> {
     let mut cargo = cargo(project);
 
-    cargo.arg("test").arg("--bin").arg(&test.bin).arg("--quiet");
+    cargo
+        .arg("test")
+        .arg("--bin")
+        .arg(&test.bin)
+        .arg("--color")
+        .arg("never")
+        .arg("--quiet");
 
     // We don't want a backtrace to dilute our snapshots (or make them instable):
     cargo.env("RUST_BACKTRACE", "0");
@@ -188,7 +194,13 @@ pub(crate) fn test(project: &Project, test: &Test, options: &Options) -> Result<
 pub(crate) fn run(project: &Project, test: &Test, options: &Options) -> Result<CargoOutput> {
     let mut cargo = cargo(project);
 
-    cargo.arg("run").arg("--bin").arg(&test.bin).arg("--quiet");
+    cargo
+        .arg("run")
+        .arg("--bin")
+        .arg(&test.bin)
+        .arg("--color")
+        .arg("never")
+        .arg("--quiet");
 
     // We don't want a backtrace to dilute our snapshots (or make them instable):
     cargo.env("RUST_BACKTRACE", "0");